+++ packages/logtape/src/formatter.ts
@@ -595,7 +595,9 @@ export interface AnsiColorFormatterOptions extends TextFormatterOptions {
     | "time"
     | "date"
     | "rfc3339"
-    | ((ts: number) => string);
+    | "none"
+    | "disabled"
+    | ((ts: number) => string | null);
 
   /**
    * The ANSI style for the timestamp.  `"dim"` is used by default.
@@ -679,12 +681,16 @@ export function getAnsiColorFormatter(
     ...options,
     format({ timestamp, level, category, message, record }): string {
       const levelColor = levelColors[record.level];
-      timestamp = `${timestampPrefix}${timestamp}${timestampSuffix}`;
+      timestamp = timestamp == null
+        ? null
+        : `${timestampPrefix}${timestamp}${timestampSuffix}`;
       level = `${levelStyle == null ? "" : ansiStyles[levelStyle]}${
         levelColor == null ? "" : ansiColors[levelColor]
       }${level}${levelStyle == null && levelColor == null ? "" : RESET}`;
       return format == null
-        ? `${timestamp} ${level} ${categoryPrefix}${category}:${categorySuffix} ${message}`
+        ? `${
+          timestamp == null ? "" : `${timestamp} `
+        }${level} ${categoryPrefix}${category}:${categorySuffix} ${message}`
         : format({
           timestamp,
           level,